pyproject.toml
[tool.poetry]
name = "python-falcon-api"
version = "0.1.0"
description = ""
authors = ["João Victor"]

[tool.poetry.dependencies]
python = "^3.8"
PyJWT = "^2.1.0"
python-dotenv = "^0.19.0"
falcon = "^3.0.1"

[tool.poetry.dev-dependencies]
conventional-commit = "^0.4.2"
autopep8 = "^1.5.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
